fix: 타입 선언 제거 및 데이터 분리
