+++ apps/web/src/components/common/profile/menu-list.tsx
@@ -1,39 +1,5 @@
 import Link from 'next/link';
-
-interface MenuItem {
-  title: string;
-  route?: string;
-  url?: string;
-}
-
-interface MenuSection {
-  id: string;
-  title: string;
-  items: MenuItem[];
-}
-
-const menuSections: MenuSection[] = [
-  {
-    id: 'setting',
-    title: '설정',
-    items: [
-      { title: '내 동네 설정', route: '/profile/location' },
-      { title: '거래 주소 관리', route: '/address' },
-      { title: '거래 계좌 관리', route: '/account' },
-    ],
-  },
-  {
-    id: 'support',
-    title: '고객 지원',
-    items: [
-      { title: '공지사항', route: '/profile/support' },
-      {
-        title: '문의하기',
-        url: 'https://docs.google.com/forms/d/e/1FAIpQLScMWkhRBarGwD0Tu7hedNc8XIRYvdJ0BYyNqKBBRyf31XIAtg/viewform?usp=header',
-      },
-    ],
-  },
-];
+import { menuSections } from '@/constants/profile';
 
 const MenuList = () => (
   <nav aria-label="마이페이지 메뉴 목록" className="px-8">
